Introduce DRS controller (#2305)
