+++ src/azul/drs.py
@@ -1,26 +1,12 @@
-from ast import (
-    literal_eval,
-)
-import base64
 from collections import (
     namedtuple,
-)
-from dataclasses import (
-    dataclass,
-    field,
-)
-from datetime import (
-    datetime,
 )
 from enum import (
     Enum,
 )
 import time
 from typing import (
-    List,
-    Mapping,
     Optional,
-    Tuple,
     Union,
 )
 
@@ -33,153 +19,34 @@ from more_itertools import (
 import requests
 
 from azul import (
-    CatalogName,
-    config,
-    dss,
     reject,
     require,
 )
-from azul.types import (
-    JSON,
-    MutableJSON,
-)
 
 
-def object_url(file_uuid: str,
-               file_version: Optional[str] = None,
-               base_url: Optional[str] = None) -> str:
-    """
-    The drs:// URL for a given DSS file UUID and version. The return value will
-    point at the bare-bones DRS data object endpoint in the web service.
-
-    :param file_uuid: the DSS file UUID of the file
-
-    :param file_version: the DSS file version of the file
-
-    :param base_url: an optional service endpoint, e.g for local test servers.
-                     If absent, the service endpoint for the current deployment
-                     will be used.
-    """
-    scheme, netloc = _endpoint(base_url)
+def drs_object_uri(host: str, path: str, **params: str) -> str:
+    assert ':' not in host, host
     return furl(scheme='drs',
-                netloc=netloc,
-                path=file_uuid,
-                args=_url_query(file_version)
+                netloc=host,
+                path=path,
+                args=params,
                 ).url
 
 
-def dos_http_object_url(catalog: CatalogName,
-                        file_uuid: str,
-                        file_version: Optional[str] = None,
-                        base_url: Optional[str] = None) -> str:
+def drs_object_url_path(object_id: str, access_id: str = None) -> str:
     """
-    The http:// or https:// URL for a given DSS file UUID and version. The
-    return value will point at the bare-bones DOS data object endpoint in the
-    web service.
-
-    :param catalog: the name of the catalog to retrieve the file from
-
-    :param file_uuid: the DSS file UUID of the file
-
-    :param file_version: the DSS file version of the file
-
-    :param base_url: an optional service endpoint, e.g for local test servers.
-                     If absent, the service endpoint for the current deployment
-                     will be used.
-    """
-    scheme, netloc = _endpoint(base_url)
-    return furl(scheme=scheme,
-                netloc=netloc,
-                path=dos_http_object_path(file_uuid),
-                query_params=dict(_url_query(file_version), catalog=catalog)
-                ).url
-
-
-def http_object_url(file_uuid: str,
-                    file_version: Optional[str] = None,
-                    base_url: Optional[str] = None,
-                    access_id: Optional[str] = None) -> str:
-    """
-    The http:// or https:// URL for a given DSS file UUID and version. The
-    return value will point at the bare-bones DRS data object endpoint in the
-    web service.
-
-    :param file_uuid: the DSS file UUID of the file
-
-    :param file_version: the optional DSS file version of the file
-
-    :param base_url: an optional service endpoint, e.g for local test servers.
-                     If absent, the service endpoint for the current deployment
-                     will be used.
-
-    :param access_id: access id will be included in the URL if this parameter is
-                      supplied
-    """
-    scheme, netloc = _endpoint(base_url)
-    return furl(scheme=scheme,
-                netloc=netloc,
-                path=http_object_path(file_uuid, access_id=access_id),
-                args=_url_query(file_version),
-                ).url
-
-
-def http_object_path(file_uuid: str, access_id: str = None) -> str:
-    """
-    >>> http_object_path('abc')
+    >>> drs_object_url_path('abc')
     '/ga4gh/drs/v1/objects/abc'
 
-    >>> http_object_path('abc', access_id='123')
+    >>> drs_object_url_path('abc', access_id='123')
     '/ga4gh/drs/v1/objects/abc/access/123'
     """
     drs_url = '/ga4gh/drs/v1/objects'
-    return '/'.join((drs_url, file_uuid, *(('access', access_id) if access_id else ())))
+    return '/'.join((drs_url, object_id, *(('access', access_id) if access_id else ())))
 
 
-def dos_http_object_path(file_uuid: str) -> str:
-    return f'/ga4gh/dos/v1/dataobjects/{file_uuid}'
-
-
-def _endpoint(base_url: Optional[str]) -> Tuple[str, str]:
-    if base_url is None:
-        base_url = config.drs_endpoint()
-    base_url = furl(base_url)
-    return base_url.scheme, base_url.netloc
-
-
-def _url_query(file_version: Optional[str]) -> Mapping[str, str]:
-    return {'version': file_version} if file_version else {}
-
-
-def encode_access_id(token_str: str, replica: str) -> str:
-    """
-    Encode a given token as an access ID using URL-safe base64 without padding.
-
-    Standard base64 pads the result with equal signs (`=`). Those would need to
-    be URL-encoded when used in the query portion of a URL:
-
-    >>> base64.urlsafe_b64encode(b"('back on boogie street', 'aws')")
-    b'KCdiYWNrIG9uIGJvb2dpZSBzdHJlZXQnLCAnYXdzJyk='
-
-    This function strips that padding. The padding is redundant as long as the
-    length of the encoded string is known at the time of decoding. With URL
-    query parameters this is always the case.
-
-    >>> encode_access_id('back on boogie street', 'aws')
-    'KCdiYWNrIG9uIGJvb2dpZSBzdHJlZXQnLCAnYXdzJyk'
-
-    >>> decode_access_id(encode_access_id('back on boogie street', 'aws'))
-    ('back on boogie street', 'aws')
-    """
-    access_id = repr((token_str, replica)).encode()
-    access_id = base64.urlsafe_b64encode(access_id)
-    return access_id.rstrip(b'=').decode()
-
-
-def decode_access_id(access_id: str) -> str:
-    token = access_id.encode('ascii')  # Base64 is a subset of ASCII
-    padding = b'=' * (-len(token) % 4)
-    token = base64.urlsafe_b64decode(token + padding)
-    return literal_eval(token.decode())
+def dos_object_url_path(object_id: str) -> str:
+    return f'/ga4gh/dos/v1/dataobjects/{object_id}'
 
 
 class AccessMethod(namedtuple('AccessMethod', 'scheme replica'), Enum):
@@ -188,72 +55,6 @@ class AccessMethod(namedtuple('AccessMethod', 'scheme replica'), Enum):
 
     def __str__(self) -> str:
         return self.name
-
-
-@dataclass
-class DRSObject:
-    """"
-    Used to build up a https://ga4gh.github.io/data-repository-service-schemas/docs/#_drsobject
-    """
-    uuid: str
-    version: Optional[str] = None
-    access_methods: List[MutableJSON] = field(default_factory=list)
-
-    def add_access_method(self,
-                          access_method: AccessMethod, *,
-                          url: Optional[str] = None,
-                          access_id: Optional[str] = None):
-        """
-        We only currently use `url_type`s of 'https' and 'gs'. Only one of `url`
-        and `access_id` should be specified.
-        """
-        assert url is None or access_id is None
-        self.access_methods.append({
-            'type': access_method.scheme,
-            **({} if access_id is None else {'access_id': access_id}),
-            **({} if url is None else {'access_url': {'url': url}}),
-        })
-
-    def to_json(self) -> JSON:
-        version = (f'&version={self.version}' if self.version is not None else '')
-        url = f'{config.dss_endpoint}/files/{self.uuid}?replica=aws' + version
-        headers = requests.head(url).headers
-        version = headers['x-dss-version']
-        if self.version is not None:
-            assert version == self.version
-        return {
-            **{
-                'checksums': [
-                    {'sha1': headers['x-dss-sha1']},
-                    {'sha-256': headers['x-dss-sha256']}
-                ],
-                'created_time': timestamp(version),
-                'id': self.uuid,
-                'self_uri': object_url(self.uuid, version),
-                'size': headers['x-dss-size'],
-                'version': version
-            },
-            'access_methods': self.access_methods
-        }
-
-
-def timestamp(version):
-    """
-    Convert a DSS version into a proper, RFC3339 compliant timestamp.
-
-    >>> timestamp('2019-08-01T211621.345939Z')
-    '2019-08-01T21:16:21.345939Z'
-
-    >>> timestamp('2019-08-01T211621:345939Z')
-    Traceback (most recent call last):
-    ...
-    ValueError: time data '2019-08-01T211621:345939Z' does not match format '%Y-%m-%dT%H%M%S.%fZ'
-    """
-    return datetime.strptime(version, dss.version_format).isoformat() + 'Z'
-
-
-def access_url(url):
-    return {'url': url}
 
 
 class DRSClient:
@@ -293,7 +94,7 @@ class DRSClient:
                f'The value {drs_uri} is not a valid hostname-based DRS URI')
         parsed.scheme = 'https'
         object_id = one(parsed.path.segments)
-        parsed.path.set(http_object_path(object_id, access_id))
+        parsed.path.set(drs_object_url_path(object_id, access_id))
         return parsed.url
 
     def _get_object(self, drs_uri: str, access_method: AccessMethod) -> str:
